Initial Version
This will check if all movie under E:\Plex\Movies has a hard link and that hard link target location is in E:\deluge\Data, if not will prompt the user if they want to delete the file. By default ANY file that does not have a hardlink will be deleted.